Add product app
In this commit i have added product as an app
to avoid conflicts that can occur in the future
since other people have similar changes,also
finished product module with tests and everything
in the previous commit is already here but other
than being on the buy_together_app i have put it
in the product app to have everything that is
related to products.
