+++ buy_together/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'buy_together_app.apps.BuytogetherappConfig',
+    'product.apps.ProductConfig',
 ]
 
 MIDDLEWARE = [
