--- buy_together_app/tests.py
@@ -1,2 +1,0 @@
-def test_temp():
-    assert 1 == 1
