+++ product/urls.py
@@ -14,8 +14,8 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.urls import path
-from buy_together_app.related_to_product import products_views
+from product import views
 
 urlpatterns = [
-    path('', products_views.index, name="product_list"),
+    path('', views.index, name="product_list"),
 ]
